requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.6.0
    # via django
attrs==22.2.0
    # via pytest
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
django==4.1.7
    # via
    #   -r requirements.in
    #   django-cas-ng
    #   django-countries-plus
    #   django-languages-plus
    #   djangorestframework
django-cas-ng==4.3.0
    # via -r requirements.in
django-countries-plus==2.1.0
    # via django-languages-plus
django-languages-plus==2.0.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
exceptiongroup==1.1.1
    # via pytest
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
lxml==4.9.2
    # via python-cas
packaging==23.0
    # via pytest
pluggy==1.0.0
    # via pytest
psycopg2==2.9.5
    # via -r requirements.in
pytest==7.2.2
    # via pytest-django
pytest-django==4.5.2
    # via -r requirements.in
python-cas==1.6.0
    # via django-cas-ng
pytz==2023.2
    # via djangorestframework
requests==2.28.2
    # via
    #   django-countries-plus
    #   python-cas
six==1.16.0
    # via python-cas
sqlparse==0.4.3
    # via django
tomli==2.0.1
    # via pytest
urllib3==1.26.15
    # via
    #   django-countries-plus
    #   requests
